move table to features
