+++ src/features/table/ChangesRecordCard.tsx
