+++ src/pages/content/pages/Folder.tsx
@@ -6,7 +6,7 @@ import CompactLayout from "@src/components/layouts/CompactLayout"
 import MainLayout from "@src/components/layouts/MainLayout"
 import WideLayout from "@src/components/layouts/WideLayout"
 import { QuickLinks } from "@src/features/quickAccess/"
-import Table from "@src/components/table/Table"
+import Table from "@src/features/table/Table"
 import { ThemeToggle } from "@src/components/ui/theme-toggle"
 import { FileMetricsTracker } from "@src/features/versionControl"
 import { FolderContent } from "@src/features/parser"
